editoast: isolate electrical profiles CLI in its own module
